backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0
python-socketio>=5.9.0
python-engineio>=4.7.0
python-dotenv>=1.0.0
pymongo>=4.6.1
python-dateutil>=2.8.2
PyJWT>=2.8.0
bcrypt>=4.1.0
APScheduler>=3.10.0
